+++ views/admin-cms-tree-type-property/_form.php
@@ -61,8 +61,8 @@ use skeeks\cms\modules\admin\widgets\Pjax;
 
 
     <?= $form->fieldRadioListBoolean($model, 'searchable') ?>
-    <?= $form->fieldRadioListBoolean($model, 'filtrable') ?>
-    <?= $form->fieldRadioListBoolean($model, 'smart_filtrable') ?>
+    <?/*= $form->fieldRadioListBoolean($model, 'filtrable') */?><!--
+    --><?/*= $form->fieldRadioListBoolean($model, 'smart_filtrable') */?>
     <?= $form->fieldRadioListBoolean($model, 'with_description') ?>
 <?= $form->fieldSetEnd(); ?>